Fix async tests + drop Python 3.5
