+++ codecov.yaml
@@ -1,0 +1,3 @@
+ignore:
+  - "tests"
+  - "setup.py"
