+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = reretry
-version = 0.11.1
+version = 0.11.2
 author = leshchenko1979
 author_email = leshchenko1979@gmail.com
 summary = An easy to use, but functional decorator for retrying on exceptions.
@@ -8,7 +8,7 @@ license = Apache License 2.0
 description_file = README.md
 long_description_content_type = text/x-rst
 home_page = https://github.com/leshchenko1979/reretry
-requires_python = >=3.5
+requires_python = >=3.6
 classifier =
     Development Status :: 4 - Beta
     Intended Audience :: Developers
@@ -16,7 +16,6 @@ classifier =
     Natural Language :: English
     Operating System :: OS Independent
     Programming Language :: Python
-    Programming Language :: Python :: 3.5
     Programming Language :: Python :: 3.6
     Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
